Removed bull board from queue module.
