+++ sources/packages/backend/apps/queue-consumers/src/constants/system-configurations.constants.ts
@@ -1,0 +1,4 @@
+/**
+ * Bull Dashboard route.
+ */
+export const BULL_BOARD_ROUTE = "admin/queues";
