+++ sources/packages/backend/apps/queue-consumers/src/queue-consumers.module.ts
@@ -78,6 +78,7 @@ import { CASSupplierIntegrationService } from "./services/cas-supplier/cas-suppl
 import { VirusScanProcessor } from "./processors/virus-scan/virus-scan.processor";
 import { CASService } from "@sims/integrations/cas/cas.service";
 import { ObjectStorageService } from "@sims/integrations/object-storage";
+import { BullBoardQueuesModule } from "./bull-board/bull-board-queues.module";
 
 // TODO: Removed ATBCResponseIntegrationScheduler in providers, the queuename from enum and the decorators of the processor as part of #2539.
 @Module({
@@ -85,6 +86,7 @@ import { ObjectStorageService } from "@sims/integrations/object-storage";
     GlobalHttpModule,
     DatabaseModule,
     QueueModule,
+    BullBoardQueuesModule,
     ZeebeModule.forRoot(),
     IER12IntegrationModule,
     ECEIntegrationModule,
